+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Don't forget to remove deprecated code on each major release!
 - Fix issues with parsing excess whitespace within `dbbackup -d "<COMMA_SEPARATED_ARGS>"`
 - Fix encryption support when using `gnupg==5.x`.
 - Resolve SQLite backup temporary file locking issues on Windows.
+- Fix MediaRestore path corruption for files containing "media" in their paths.
 
 ### Security